refactor: consolidate TypeScript types and update imports
- Move types from @types/common.ts to types.ts
- Remove @types/global.d.ts
- Update all type imports to use new types.ts
- Add JSDoc documentation to type definitions
- Align with Lodash integration updates
