--- src/@types/global.d.ts
@@ -1,2 +1,0 @@
-declare const __RSUITE_CLASSNAME_PREFIX__: boolean;
-declare const __DEV__: boolean;
