+++ src/hooks/useTableData.ts
@@ -4,7 +4,7 @@ import flattenData from '../utils/flattenData';
 import findAllParents from '../utils/findAllParents';
 import shouldShowRowByExpanded from '../utils/shouldShowRowByExpanded';
 import { EXPANDED_KEY, TREE_DEPTH } from '../constants';
-import type { RowKeyType, RowDataType } from '../@types/common';
+import type { RowKeyType, RowDataType } from '../types';
 /**
  * Filter those expanded nodes.
  * @param data
