+++ src/index.ts
@@ -4,16 +4,9 @@ export { default as Cell } from './Cell';
 export { default as HeaderCell } from './HeaderCell';
 export { default as ColumnGroup } from './ColumnGroup';
 
-export type { TableLocaleType } from './@types/common';
 export type { TableProps, TableInstance } from './Table';
 export type { ColumnProps } from './Column';
 export type { CellProps } from './Cell';
 export type { HeaderCellProps } from './HeaderCell';
 export type { ColumnGroupProps } from './ColumnGroup';
-export type {
-  StandardProps,
-  SortType,
-  RowDataType,
-  RowKeyType,
-  TableSizeChangeEventName
-} from './@types/common';
+export * from './types';
